feat(Backend): localisation
Now, when a company is created, you have to give it's address. With that
address, the server will get the lat/lon of the position and store it to
the DB. Thanks to PostGIS and sequelize, the new resolver
"getCompaniesByDistance" allow now to get the companies based on a
geographic point in space given in argument.